feat: complete macOS/ARM64 port — compiler, runtime, tests, benchmarks
Port the TML compiler and runtime to macOS/ARM64 (Apple Silicon).

Compiler C++ (Phase 1): Clang compatibility fixes — template disambiguation,
use-after-move UB, [[maybe_unused]], static_cast, platform guards, safe casts.

C Runtime (Phase 2): POSIX compat layer (compat.h), signal handlers with
sigaction+siglongjmp, pthread-based threading with custom stacks, malloc_size
for Apple, clock_gettime for time_ns.

Standard Library (Phase 3): Platform-specific socket constants for macOS,
RawSocket::as_raw() method, FFI bindings verified for POSIX.

Build System (Phase 4): build.sh with Homebrew LLVM auto-detection, system
linker on macOS (no LLD), OpenSSL/zstd/brotli Homebrew linking, MCP server
Unix path resolution.

Test Runner (Phase 5): Fixed 10 runtime failures —
- 5 str CRASH (generic replace[T] misresolution)
- 1 ops_misc SIGSEGV (ManuallyDrop API unification)
- 1 unicode_char PANIC (module name collision)
- 3 error PANIC (when expression use-after-free in pending_str_temps_)
- 1 zstd_dict_train SIGILL (double-dereference in List[Buffer] FFI)

Result: ~5,921 tests passing, 0 runtime crashes on macOS ARM64.

Benchmarks (Phase 6): TCP/UDP request round-trip benchmarks across TML, Rust,
Go, and Node.js. TML wins bind latency (7.7µs TCP, 6.5µs UDP), competitive
with Rust on request latency.

Documentation (Phase 7): Updated CLAUDE.md and README.md with macOS build
instructions, platform support table, and Homebrew prerequisites.

+++ compiler/src/hir/serializer/serialize_utils.cpp
@@ -88,7 +88,7 @@ public:
     }
 
     void update(const std::string& str) {
-        update(str.size());
+        update(static_cast<uint64_t>(str.size()));
         update(str.data(), str.size());
     }
